Politicas para administrador

+++ Practica1/app/Http/Controllers/Personas.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Policies\PersonaPolicy;
 use App\Models\Persona;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -17,13 +18,20 @@ class Personas extends Controller {
      *
      * @return Response
      */
-    public function index(/*$lang*/) {
+    public function index(/* $lang */) {
         //App::setLocale($lang);
         if (Gate::allows('pais-prueba')) {
             //$paises = Paises::all();
             //$paises = Paises::simplePaginate(15);
             $personas = Persona::all();
-            return view('Personas.index')->with('personas', $personas);
+            /* Forma de enviar parametros con with, primero el nombre de
+             * la futura variable y después el valor que envias
+             *  return view('Personas.index')->with('personas', $personas);
+             */
+            return view('Personas.index', [
+                'personas' => $personas,
+                'policy' => new Persona
+            ]);
         }
         Abort(403);
     }
@@ -34,6 +42,10 @@ class Personas extends Controller {
      * @return Response
      */
     public function create() {
+        /* Autorizar el acceso a la web a solo los
+         * que sean admin a través de Policies */
+        $this->authorize('create', new Persona);
+
         return view('Personas.create');
     }
 
@@ -44,6 +56,9 @@ class Personas extends Controller {
      * @return Response
      */
     public function store(Request $request) {
+
+        $this->authorize('create', new Persona);
+
         Persona::create($request->all());
         return redirect('personas')->with('flash_message', 'Persona añadida');
     }
@@ -66,6 +81,8 @@ class Personas extends Controller {
      * @return Response
      */
     public function edit($id) {
+        $this->authorize('update', new Persona);
+
         $persona = Persona::find($id);
         return view('Personas.edit')->with('personas', $persona);
     }
@@ -78,6 +95,8 @@ class Personas extends Controller {
      * @return Response
      */
     public function update(Request $request, $id) {
+        $this->authorize('update', new Persona);
+
         $persona = Persona::find($id);
         $persona->update($request->all());
         return redirect('personas')->with('flash_message', 'Persona actualizada');
@@ -90,6 +109,8 @@ class Personas extends Controller {
      * @return Response
      */
     public function destroy($id) {
+        $this->authorize('delete', new Persona);
+
         Persona::destroy($id);
         return redirect('personas')->with('flash_message', 'Persona borrada');
     }
